Initial support for unary operators
- (negation, prefix), + (positive, prefix, but redundant because values are
assumed positive), and ! (factorial, postfix) unary operators are now partially
supported. The stack_item struct was expanded to have a "flags" field, which is
generic, but for now is just used to store whether the operator is unary or not.
That's used later on when looping through the stack to determine how an operator
should be applied.

One bug is that in certain cases when an unary operator is used within parens,
it'll fail with a syntax error.