Store doubles without precision loss
Since the stack stores all values as strings, doubles need to be converted to
strings before they can be pushed to the stack. printf-style %g wasn't cutting
it since it was very imprecise. Luckily there's an %a which stores it as a
hexadecimal string with no loss of precision.

However, %a isn't human-readable, so I have to explicitly convert it to decimal
for the final output. I do this using %f (because %g limits itself based on
total significant figures rather than decimal places, which I don't want).
trim_double() was created to avoid ugly output like "2.000000".