Fix order of operations inside parens
This is actually a surprisingly simple fix. Remove the !paren_depth check so
that compare_operators() is still used inside parens, then add "(" to the end of
the op order array (so it's the lowest precedence). The apply_operator() loop
when a ")" is encountered is kept to clean up and merge the parenthetical into
the main expression, but by then all the order of operations logic is taken care
of already.