Better error output
All the error messages have been completely rewritten to be more consistent,
succinct and clear. Column number and character have been removed from the
messages, instead replaced with an excerpt of the expression (or the full
expression if it's short enough) with a caret (^) underneath it pointing at the
column where the error was encountered.

I think it looks pretty nice. It would be cool to have color output, but that's
a bit complicated, so this will do fine for now.

Also, ERROR_UNRECOGNIZED now exits the program instead of continuing on (which
would make more sense for a warning than an error).

One minor regression is that for errors encountered on the stack instead of in
the string loop, the problem character is no longer printed, and since the
column number is unknown for those errors, no info about the location of the
error is displayed. That'll be fixed when I store the column number in the flags
field of the stack.