Add unit tests using CUnit
Comes with two complementary test failures. :( But I'm glad I have this set up
now to avoid regressions in the future. :D

Also: some shuffling around of #includes, wanted to make it easier on linking
new files with shunting-yard, so some includes are made from shunting-yard.h
itself now. I'm not 100% sure if this is good practice, but it seems like the
best solution right now.