Basic function support
Function names are considered operands, and when encountering parentheses, the
program checks if the previous token is an operand. If it is, it tries to apply
that function. Only one argument is allowed for now (no commas). Currently
supported: sqrt, cos, sin, tan.

This can be used to fairly easily add "2(3)"-style multiplication next.