Return error for functions without arguments
This fixes something like "abs()" returning a segfault (due to stack underflow).

One weird side effect is that the argument is check is made first, _then_ the
check for undefined functions. So "foo()" will return "functions need an
argument", but when you fix it to "foo(2)" it'll say "undefined function". Minor
waste of time, but definitely should not be an issue at all during normal usage,
so it's not really worth the time to fix (the code is cleaner without a fix).

(Also: fixed unfreed memory when encountering a function error (the goto skipped
over the call to free()))