Add equality operator
The equality operator ("=" - using this instead of "==" to match normal math) is
now supported. This _would've_ been a simple patch, but extra code was needed to
allow more than one "=" in an expression (such as "2=2=2"). Substituting each
equality comparison with 1 or 0 doesn't work, because that works like:
    1. 2=2=2
    2. (2=2)=2
    3. (1)=2
    4. false
Simplifying that by just replacing it with the result value works for every case
_except_ 0=0 (since 0 is the same as false). So the best solution I could think
of involves using the stack flags to set whether a comparison is true or not
(FLAG_BOOL_TRUE in config.h). Using that value instead of interpreting the
result itself as a bool works in all cases, although I'm a bit annoyed that such
a complex solution was needed for only one case ("0=0").