Use struct to hold operator info
Instead of an array of strings for operators, there's now an array of structs
that specify operator character, precedence, and type. This looks a bit nicer in
code and is more scalable, I think.

The new type, op_t, will be used to pass operators to different functions where
it makes sense. This is cool because it means there just needs to be one op_t
argument instead of two (char op, bool unary).